+++ core/correctors/rgb-channel-corrector.h
@@ -16,14 +16,14 @@ namespace Corrector {
 		double greenFactor() const;
 		double blueFactor() const;
 
-		void setRedFactor(const double &factor);
-		void setGreenFactor(const double &factor);
-		void setBlueFactor(const double &factor);
+		void setRedFactor(const float &factor);
+		void setGreenFactor(const float &factor);
+		void setBlueFactor(const float &factor);
 
 	private:
-		std::atomic<double> m_rfactor;
-		std::atomic<double> m_gfactor;
-		std::atomic<double> m_bfactor;
+		std::atomic<float> m_rfactor;
+		std::atomic<float> m_gfactor;
+		std::atomic<float> m_bfactor;
 	};
 
 }
